+++ src/app/shared/shared.module.ts
@@ -7,6 +7,8 @@ import { OfferTileComponent } from 'src/app/shared/components/offer-tile/offer-t
 import { ErrorBackTextComponent} from 'src/app/shared/components/error-back-text/error-back-text.component';
 import { LoadingSpinnerComponent } from 'src/app/shared/components/loading-spinner/loading-spinner.component';
 import { DrawableCanvasComponent } from 'src/app/shared/components/drawable-canvas/drawable-canvas.component';
+import { InfoBannerComponent } from 'src/app/shared/components/info-banner/info-banner.component';
+import { ConsentVideoComponent } from 'src/app/shared/components/consent-video/consent-video.component';
 
 import {
   FaIconLibrary,
@@ -35,7 +37,9 @@ declarations: [
   OfferTileComponent,
   ErrorBackTextComponent,
   LoadingSpinnerComponent,
-  DrawableCanvasComponent
+  DrawableCanvasComponent,
+  InfoBannerComponent,
+  ConsentVideoComponent
 ],
 imports: [
   CommonModule,
@@ -51,7 +55,9 @@ exports: [
   OfferTileComponent,
   ErrorBackTextComponent,
   LoadingSpinnerComponent,
-  DrawableCanvasComponent
+  DrawableCanvasComponent,
+  InfoBannerComponent,
+  ConsentVideoComponent
 ]
 })