+++ src/app/sites/landing/components/banner-top/banner-top.component.ts
@@ -2,8 +2,7 @@ import { Component, OnInit } from '@angular/core';
 
 @Component({
   selector: 'app-banner-top',
-  templateUrl: './banner-top.component.html',
-  styleUrls: ['./banner-top.component.scss']
+  templateUrl: './banner-top.component.html'
 })
 export class PreviewInfoBannerComponent implements OnInit {